#136 refactor api integration
