+++ modules/gui/package.json
@@ -43,12 +43,14 @@
 	"dependencies": {
 		"@tauri-apps/api": "^1.2.0",
 		"@types/bcryptjs": "^2.4.2",
+		"@types/url-join": "^4.0.1",
 		"bcryptjs": "^2.4.3",
 		"buffer": "^6.0.3",
 		"fuse.js": "^6.6.2",
 		"lodash": "^4.17.21",
 		"lorem-ipsum": "^2.0.8",
-		"svelte-watch-resize": "^1.0.3"
+		"svelte-watch-resize": "^1.0.3",
+		"url-join": "^5.0.0"
 	},
 	"pnpm": {
 		"onlyBuiltDependencies": [
